feat: idempotent seedVotables
